restructured code based on feedback
