+++ lib/cubits/authentication/authentication_cubit.dart
@@ -7,12 +7,13 @@ part 'authentication_state.dart';
 
 enum AuthStatus { unknown, authenticated, unauthenticated }
 
+//We might consider changing this back to a bloc if we want different events to
+//trigger a logout (for instance, when the user requests logs out themselves
+//vs the user's token expires and fails to renew).
 class AuthenticationCubit extends Cubit<AuthenticationState> {
   final SecureStorage _storage;
 
-  AuthenticationCubit(this._storage) : super(const AuthenticationState._()) {
-    appStarted();
-  }
+  AuthenticationCubit(this._storage) : super(const AuthenticationState._());
 
   Future<void> appStarted() async {
     final authenticatedUser = await _storage.getAuthenticatedUser();
