+++ lib/cubits/environment_type/environment_type_cubit.dart
@@ -1,9 +1,7 @@
 import 'package:bloc/bloc.dart';
-import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/data/repositories/v1/app_config_repository.dart';
 import 'package:coffeecard/service_locator.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 
 part 'environment_type_state.dart';
 
@@ -30,26 +28,7 @@ class EnvironmentTypeCubit extends Cubit<EnvironmentTypeState> {
     emit(EnvironmentTypeState(status: isConnectedToTestDB));
   }
 
-  Future<void> addTestDBWidget(BuildContext context) async {
-    //Delay is required to trigger after build
-    Future.delayed(Duration.zero, () {
-      final entry = OverlayEntry(
-        builder: (context) => IgnorePointer(
-          child: Align(
-            alignment: Alignment.bottomCenter,
-            child: SizedBox(
-              height: 100,
-              child: Text(
-                Strings.testDBString,
-                style: TextStyle(color: Colors.red[900], fontSize: 20),
-              ),
-            ),
-          ),
-        ),
-      );
-      final overlay = Overlay.of(context);
-      overlay?.insert(entry);
-    });
+  Future<void> signalWidgetAdded() async {
     EnvironmentTypeState.widgetAdded = true;
     emit(EnvironmentTypeState(status: state.status));
   }
