--- lib/widgets/pages/tickets_page/tickets_page.dart
@@ -1,39 +1,0 @@
-import 'package:coffeecard/base/strings.dart';
-import 'package:coffeecard/base/style/colors.dart';
-import 'package:coffeecard/base/style/text_styles.dart';
-import 'package:coffeecard/cubits/analog_closed_popup/analog_closed_popup_cubit.dart';
-import 'package:coffeecard/cubits/tickets_page/tickets_cubit.dart';
-import 'package:coffeecard/errors/match_case_incomplete_exception.dart';
-import 'package:coffeecard/widgets/components/coffee_card.dart';
-import 'package:coffeecard/widgets/components/helpers/tappable.dart';
-import 'package:coffeecard/widgets/components/rounded_button.dart';
-import 'package:coffeecard/widgets/pages/buy_one_drink_page.dart';
-import 'package:coffeecard/widgets/pages/buy_other_page.dart';
-import 'package:coffeecard/widgets/pages/buy_tickets_page.dart';
-import 'package:coffeecard/widgets/pages/redeem_voucher_page.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-
-part 'analog_closed_popup.dart';
-part 'tickets_section.dart';
-part 'shop_section.dart';
-
-class TicketsPage extends StatefulWidget {
-  const TicketsPage({Key? key}) : super(key: key);
-
-  @override
-  _TicketsPageState createState() => _TicketsPageState();
-}
-
-class _TicketsPageState extends State<TicketsPage> {
-  @override
-  Widget build(BuildContext context) {
-    return ListView(
-      children: const [
-        AnalogClosedPopup(),
-        TicketSection(),
-        ShopSection(),
-      ],
-    );
-  }
-}
